% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_class.R
\name{as.pep}
\alias{as.pep}
\title{Coerce to PEP725 Data Object}
\usage{
as.pep(x, ...)
}
\arguments{
\item{x}{A data.frame or data.table to coerce.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An object of class \code{pep}.
}
\description{
Coerce to PEP725 Data Object
}
\author{
Matthias Templ
}
