% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_daylength.R
\name{calc_max_daylength}
\alias{calc_max_daylength}
\title{Maximum Daylength at a Latitude}
\usage{
calc_max_daylength(lat)
}
\arguments{
\item{lat}{Numeric. Latitude in decimal degrees.}
}
\value{
Numeric. Maximum daylength in hours.
}
\description{
Calculates the maximum possible daylength (at summer solstice) for
a given latitude.
}
\examples{
# Maximum daylength at different latitudes
calc_max_daylength(c(0, 30, 45, 60, 66.5))

}
\author{
Matthias Templ
}
