% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_class.R
\name{is.pep}
\alias{is.pep}
\title{Test if Object is a PEP725 Data Object}
\usage{
is.pep(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Logical; \code{TRUE} if \code{x} is of class \code{pep}.
}
\description{
Test if Object is a PEP725 Data Object
}
\examples{
\dontrun{
is.pep(pep)
is.pep(data.frame(x = 1))
}

}
\author{
Matthias Templ
}
