% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{meteoSwiss}
\alias{meteoSwiss}
\title{Historical Harvest and Heading Dates from MeteoSwiss}
\format{
A data.table with 30 rows and 8 variables:
\describe{
\item{yearHd}{Year of heading observation (integer)}
\item{monthHd}{Month of heading (integer)}
\item{dayHd}{Day of heading (integer)}
\item{DOYHd}{Day of year for heading (integer)}
\item{yearHv}{Year of harvest observation (integer)}
\item{monthHv}{Month of harvest (integer)}
\item{dayHv}{Day of harvest (integer)}
\item{DOYHv}{Day of year for harvest (integer)}
}
}
\source{
MeteoSwiss (Bundesamt für Meteorologie und Klimatologie), via Agroscope archives
}
\usage{
data(meteoSwiss)
}
\description{
This dataset contains manually observed heading and harvest dates for a crop (likely winter wheat) near Nyon/Changins, Switzerland, as archived by MeteoSwiss. The data includes day-of-year (DOY), year, and calendar dates for both phenophases over a 30-year period.
}
\details{
These observational records are used to validate phenological models and assess long-term changes in crop development stages under historical climate. The data originates from the MeteoSwiss station network near Changins, which is part of Switzerland’s official agricultural phenology monitoring.
}
\examples{
data(meteoSwiss)
plot(meteoSwiss$yearHd, meteoSwiss$DOYHd, type = "b", ylab = "Heading DOY", xlab = "Year")
plot(meteoSwiss$yearHv, meteoSwiss$DOYHv, type = "b", ylab = "Harvest DOY", xlab = "Year")
}
\keyword{datasets}
