% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_class.R
\name{plot.pep_coverage}
\alias{plot.pep_coverage}
\title{Plot Method for PEP Coverage}
\usage{
\method{plot}{pep_coverage}(x, ...)
}
\arguments{
\item{x}{A \code{pep_coverage} object.}

\item{...}{Additional arguments (unused).}
}
\value{
A ggplot object (or list of ggplot objects).
}
\description{
Plot Method for PEP Coverage
}
\author{
Matthias Templ
}
