% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_combine.R
\name{plot.pheno_combined}
\alias{plot.pheno_combined}
\title{Plot Method for Combined Time Series}
\usage{
\method{plot}{pheno_combined}(x, type = c("series", "stations", "residuals"), group = NULL, ...)
}
\arguments{
\item{x}{A \code{pheno_combined} object}

\item{type}{Character. "series" for combined time series, "stations" for
station effects, "residuals" for residual distribution.}

\item{group}{For grouped analysis, which group to plot.}

\item{...}{Additional arguments}
}
\value{
A ggplot object (invisibly)
}
\description{
Plot Method for Combined Time Series
}
