% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_gradient.R
\name{plot.pheno_gradient}
\alias{plot.pheno_gradient}
\title{Plot Method for Phenological Gradient Analysis}
\usage{
\method{plot}{pheno_gradient}(x, ...)
}
\arguments{
\item{x}{A \code{pheno_gradient} object}

\item{...}{Additional arguments passed to ggplot}
}
\value{
A ggplot object
}
\description{
Plot Method for Phenological Gradient Analysis
}
\author{
Matthias Templ
}
