% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_synchrony.R
\name{plot.pheno_synchrony}
\alias{plot.pheno_synchrony}
\title{Plot Method for Phenological Synchrony Analysis}
\usage{
\method{plot}{pheno_synchrony}(x, region = NULL, ...)
}
\arguments{
\item{x}{A \code{pheno_synchrony} object}

\item{region}{Optional region to plot (for multi-region analyses)}

\item{...}{Additional arguments passed to ggplot}
}
\value{
A ggplot object
}
\description{
Creates a time series plot of synchrony (SD) over years.
}
\author{
Matthias Templ
}
