% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_completeness.R
\name{print.pep_completeness}
\alias{print.pep_completeness}
\title{Print Method for Completeness Assessment}
\usage{
\method{print}{pep_completeness}(x, n = 15, ...)
}
\arguments{
\item{x}{A \code{pep_completeness} object}

\item{n}{Number of rows to display}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for Completeness Assessment
}
\author{
Matthias Templ
}
