% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_class.R
\name{print.pep_coverage}
\alias{print.pep_coverage}
\title{Print Method for PEP Coverage}
\usage{
\method{print}{pep_coverage}(x, ...)
}
\arguments{
\item{x}{A \code{pep_coverage} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for PEP Coverage
}
\author{
Matthias Templ
}
