% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_flag_outliers.R
\name{print.pep_outliers}
\alias{print.pep_outliers}
\title{Print Method for Outlier Detection Results}
\usage{
\method{print}{pep_outliers}(x, ...)
}
\arguments{
\item{x}{A \code{pep_outliers} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for Outlier Detection Results
}
