% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_quality.R
\name{print.pep_quality}
\alias{print.pep_quality}
\title{Print Method for Data Quality Assessment}
\usage{
\method{print}{pep_quality}(x, n = 10, ...)
}
\arguments{
\item{x}{A \code{pep_quality} object}

\item{n}{Number of rows to display}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for Data Quality Assessment
}
\author{
Matthias Templ
}
