% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phases.R
\name{print.phase_check}
\alias{print.phase_check}
\title{Print Method for Phase Check Results}
\usage{
\method{print}{phase_check}(x, ...)
}
\arguments{
\item{x}{A \code{phase_check} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for Phase Check Results
}
\author{
Matthias Templ
}
