% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phases.R
\name{print.phase_check_multi}
\alias{print.phase_check_multi}
\title{Print Method for Multi-Species Phase Check}
\usage{
\method{print}{phase_check_multi}(x, ...)
}
\arguments{
\item{x}{A \code{phase_check_multi} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for Multi-Species Phase Check
}
\author{
Matthias Templ
}
