% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_anomaly.R
\name{print.pheno_anomaly}
\alias{print.pheno_anomaly}
\title{Print Method for Phenological Anomalies}
\usage{
\method{print}{pheno_anomaly}(x, n = 10, ...)
}
\arguments{
\item{x}{A \code{pheno_anomaly} object}

\item{n}{Number of rows to display}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for Phenological Anomalies
}
\author{
Matthias Templ
}
