% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_combine.R
\name{print.pheno_combined}
\alias{print.pheno_combined}
\title{Print Method for Combined Time Series}
\usage{
\method{print}{pheno_combined}(x, ...)
}
\arguments{
\item{x}{A \code{pheno_combined} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for Combined Time Series
}
