% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_gradient.R
\name{print.pheno_gradient}
\alias{print.pheno_gradient}
\title{Print Method for Phenological Gradient Analysis}
\usage{
\method{print}{pheno_gradient}(x, ...)
}
\arguments{
\item{x}{A \code{pheno_gradient} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for Phenological Gradient Analysis
}
\author{
Matthias Templ
}
