% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_normals.R
\name{print.pheno_normals}
\alias{print.pheno_normals}
\title{Print Method for Phenological Normals}
\usage{
\method{print}{pheno_normals}(x, n = 10, ...)
}
\arguments{
\item{x}{A \code{pheno_normals} object}

\item{n}{Number of rows to display}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for Phenological Normals
}
\author{
Matthias Templ
}
