% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_pls.R
\name{print.pheno_pls}
\alias{print.pheno_pls}
\title{Print Method for PLS Phenology Results}
\usage{
\method{print}{pheno_pls}(x, ...)
}
\arguments{
\item{x}{A \code{pheno_pls} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for PLS Phenology Results
}
