% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_trend_turning.R
\name{print.pheno_turning}
\alias{print.pheno_turning}
\title{Print Method for Trend Turning Analysis}
\usage{
\method{print}{pheno_turning}(x, ...)
}
\arguments{
\item{x}{A \code{pheno_turning} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for Trend Turning Analysis
}
