% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_phenology.R
\name{print.pls_phenology}
\alias{print.pls_phenology}
\title{Print Method for PLS Phenology Results}
\usage{
\method{print}{pls_phenology}(x, ...)
}
\arguments{
\item{x}{A \code{pls_phenology} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for PLS Phenology Results
}
