% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_second_events.R
\name{print.second_events}
\alias{print.second_events}
\title{Print Method for Second Events Detection}
\usage{
\method{print}{second_events}(x, ...)
}
\arguments{
\item{x}{A \code{second_events} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print Method for Second Events Detection
}
