% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_class.R
\name{[.pep}
\alias{[.pep}
\title{Subset PEP725 Data While Preserving Class}
\usage{
\method{[}{pep}(x, ...)
}
\arguments{
\item{x}{A \code{pep} object.}

\item{...}{Arguments passed to the data.table subset method.}
}
\value{
A \code{pep} object if structure is preserved, otherwise a data.table.
}
\description{
Ensures that subsetting operations return a \code{pep} object
when the result still has the required structure.
}
\author{
Matthias Templ
}
