% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_outliers.R
\name{summary.pep_outliers}
\alias{summary.pep_outliers}
\title{Summary Method for Outlier Detection Results}
\usage{
\method{summary}{pep_outliers}(object, ...)
}
\arguments{
\item{object}{A \code{pep_outliers} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns a summary data.table
}
\description{
Summary Method for Outlier Detection Results
}
