% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_quality.R
\name{summary.pep_quality}
\alias{summary.pep_quality}
\title{Summary Method for Data Quality Assessment}
\usage{
\method{summary}{pep_quality}(object, ...)
}
\arguments{
\item{object}{A \code{pep_quality} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns a summary list
}
\description{
Summary Method for Data Quality Assessment
}
\author{
Matthias Templ
}
