% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_anomaly.R
\name{summary.pheno_anomaly}
\alias{summary.pheno_anomaly}
\title{Summary Method for Phenological Anomalies}
\usage{
\method{summary}{pheno_anomaly}(object, ...)
}
\arguments{
\item{object}{A \code{pheno_anomaly} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns a summary list
}
\description{
Summary Method for Phenological Anomalies
}
\author{
Matthias Templ
}
