% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_normals.R
\name{summary.pheno_normals}
\alias{summary.pheno_normals}
\title{Summary Method for Phenological Normals}
\usage{
\method{summary}{pheno_normals}(object, ...)
}
\arguments{
\item{object}{A \code{pheno_normals} object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns a summary list
}
\description{
Summary Method for Phenological Normals
}
\author{
Matthias Templ
}
