% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_second_events.R
\name{summary.second_events}
\alias{summary.second_events}
\title{Summary Method for Second Events Detection}
\usage{
\method{summary}{second_events}(object, ...)
}
\arguments{
\item{object}{A \code{second_events} object}

\item{...}{Additional arguments (unused)}
}
\value{
The summary data.table (invisibly)
}
\description{
Summary Method for Second Events Detection
}
