% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_trend_turning.R
\name{kendall_tau}
\alias{kendall_tau}
\title{Mann-Kendall Z-Statistic (deprecated alias for \code{mann_kendall_z})}
\usage{
kendall_tau(x)
}
\arguments{
\item{x}{Numeric vector (time series assumed to be equidistant).
\code{NA} values are silently removed.}
}
\value{
Numeric. Identical output shape to \code{mann_kendall_z()}.
}
\description{
This function was misnamed in pep725 <= 1.0.2: it returned the
Mann-Kendall Z-statistic, not Kendall's \eqn{\tau = S / (n(n-1)/2)}.
Please call \code{\link{mann_kendall_z}} instead for identical (and
better: tie-corrected, continuity-corrected) behaviour.
}
\details{
Note that this function is now a thin wrapper and will continue to
return the Z-statistic until a future major release.
}
\examples{
suppressWarnings(kendall_tau(c(120, 118, 115, 112, 110, 108, 105)))

}
\seealso{
\code{\link{mann_kendall_z}}
}
\author{
Matthias Templ
}
