% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_trend_turning.R
\name{kendall_tau}
\alias{kendall_tau}
\title{Kendall's Normalized Tau}
\usage{
kendall_tau(x)
}
\arguments{
\item{x}{Numeric vector (time series assumed to be equidistant)}
}
\value{
Numeric. Kendall's normalized tau statistic.
Values near +1 indicate increasing trend, -1 decreasing trend,
0 no trend. Approximately standard normal under no-trend null.
}
\description{
Computes Kendall's normalized tau statistic for a time series,
providing a non-parametric measure of monotonic trend.
}
\examples{
# Decreasing trend (earlier phenology)
kendall_tau(c(120, 118, 115, 112, 110, 108, 105))

# No clear trend
kendall_tau(c(120, 115, 122, 118, 121, 116, 119))

}
\author{
Matthias Templ
}
